Reorganized Conway conformance tests
