+++ eras/conway/impl/cardano-ledger-conway.cabal
@@ -114,6 +114,15 @@ library testlib
         Test.Cardano.Ledger.Conway.Arbitrary
         Test.Cardano.Ledger.Conway.Binary.Cddl
         Test.Cardano.Ledger.Conway.Binary.RoundTrip
+        Test.Cardano.Ledger.Conway.Constrained.Instances
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Cert
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Deleg
+        Test.Cardano.Ledger.Conway.Constrained.Spec.GovCert
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Gov
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Ledger
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Pool
+        Test.Cardano.Ledger.Conway.Constrained.Spec.PParams
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Utxo
         Test.Cardano.Ledger.Conway.ImpTest
         Test.Cardano.Ledger.Conway.Imp
         Test.Cardano.Ledger.Conway.Imp.EpochSpec
@@ -125,6 +134,7 @@ library testlib
         Test.Cardano.Ledger.Conway.Imp.RatifySpec
         Test.Cardano.Ledger.Conway.Proposals
         Test.Cardano.Ledger.Conway.TreeDiff
+
 
     visibility:       public
     hs-source-dirs:   testlib
@@ -140,23 +150,32 @@ library testlib
         bytestring,
         cardano-data:{cardano-data, testlib},
         containers,
+        cryptonite,
+        constrained-generators,
         plutus-ledger-api,
         deepseq,
         microlens,
         cardano-crypto-class,
+        cardano-crypto-wrapper,
         cardano-ledger-allegra:{cardano-ledger-allegra, testlib},
         cardano-ledger-alonzo:{cardano-ledger-alonzo, testlib},
         cardano-ledger-binary,
         cardano-ledger-babbage:{cardano-ledger-babbage, testlib},
+        cardano-ledger-byron,
         cardano-ledger-conway,
         cardano-ledger-core:{cardano-ledger-core, testlib},
         cardano-ledger-mary,
         cardano-ledger-shelley,
         cardano-strict-containers,
+        cardano-slotting:{cardano-slotting, testlib},
         data-default-class,
         generic-random,
         mtl,
         text,
+        vector-map,
+        QuickCheck,
+        random,
+        FailT,
         small-steps >=1.1
 
 test-suite tests
