+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Constrained/Spec/PParams.hs
@@ -1,15 +1,17 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE RecordWildCards #-}
 
-module Test.Cardano.Ledger.Constrained.V2.Conway.PParams where
+module Test.Cardano.Ledger.Conway.Constrained.Spec.PParams where
 
-import Cardano.Ledger.Api
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Conway.PParams
 
 import Constrained
 
-import Test.Cardano.Ledger.Constrained.V2.Conway
+import Cardano.Ledger.Conway (ConwayEra)
+import Cardano.Ledger.Conway.Core
+import Cardano.Ledger.Crypto (StandardCrypto)
+import Test.Cardano.Ledger.Conway.Constrained.Instances (IsConwayUniv)
 
 pparamsSpec :: IsConwayUniv fn => Spec fn (PParams (ConwayEra StandardCrypto))
 pparamsSpec =
