+++ libs/cardano-ledger-conformance/cardano-ledger-conformance.cabal
@@ -21,7 +21,8 @@ flag asserts
 library
     exposed-modules:  Test.Cardano.Ledger.Conformance
     hs-source-dirs:   src
-    other-modules:    Test.Cardano.Ledger.Conformance.Orphans
+    other-modules:
+      Test.Cardano.Ledger.Conformance.Orphans
     default-language: Haskell2010
     ghc-options:
         -Wall -Wcompat -Wincomplete-record-updates
@@ -30,26 +31,7 @@ library
 
     build-depends:
         base >=4.14 && <5,
-        bytestring,
-        cardano-strict-containers,
-        microlens,
-        mtl,
-        bytestring,
-        data-default-class,
-        cardano-ledger-binary,
-        cardano-ledger-core,
-        cardano-ledger-mary,
-        cardano-ledger-shelley,
-        cardano-ledger-alonzo,
-        cardano-ledger-babbage,
-        cardano-ledger-conway:{cardano-ledger-conway, testlib},
-        cardano-ledger-executable-spec,
-        cardano-crypto-class,
-        cardano-ledger-core:testlib,
-        cardano-ledger-test,
-        containers,
-        small-steps >=1.1,
-        text
+        cardano-ledger-executable-spec
 
     if !impl(ghc >=9.2)
         ghc-options: -Wno-incomplete-patterns
@@ -69,5 +51,3 @@ test-suite tests
 
     build-depends:
         base >=4.14 && <5,
-        cardano-ledger-conformance,
-        cardano-ledger-core:testlib
