+++ libs/cardano-ledger-conformance/test/Main.hs
@@ -1,10 +1,4 @@
 module Main where
 
-import qualified Test.Cardano.Ledger.Conformance as Conformance
-import Test.Cardano.Ledger.Imp.Common (describe, ledgerTestMain)
-
 main :: IO ()
-main =
-  ledgerTestMain $
-    describe "Conformance" $ do
-      Conformance.spec
+main = pure () -- TODO remove this module?
