+++ libs/cardano-ledger-test/bench/Bench/Constrained/STS.hs
@@ -14,8 +14,13 @@ import Cardano.Ledger.Crypto
 import Constrained
 import Constrained.Bench
 import Criterion
-import Test.Cardano.Ledger.Constrained.V2.Conway
-import Test.Cardano.Ledger.Constrained.V2.Conway.GOV
+import Test.Cardano.Ledger.Conway.Constrained.Instances
+import Test.Cardano.Ledger.Conway.Constrained.Spec.Gov (
+  allGASInTree,
+  govEnvSpec,
+  govProposalsSpec,
+  wellFormedChildren,
+ )
 
 govEnv :: GovEnv (ConwayEra StandardCrypto)
 govEnv = genFromSpecWithSeed 10 30 (govEnvSpec @ConwayFn)
